checkpoint: extracts "Version1/Image.ts"
