+++ src/library/ShimmedStabilityAIClient/definition.ts
@@ -1,51 +1,12 @@
-import type {
-  Image as StabilityAI_TextToImage_Pipeline_Output,
-} from 'stabilityai-client-typescript/models/components';
-
-import type {
-  GenerateFromTextRequest,
-  GenerateFromTextResponse,
-} from 'stabilityai-client-typescript/models/operations';
-
 import HTTP from '@/library/HTTP';
-import Shortfin from '@/library/Shortfin';
 
 import {
   URLOrigin,
 } from '@/library/URLComponent';
 
-import toShortfin from './toShortfin';
-
-class ShimmedStabilityAIClient_Version1_Image
-  extends HTTP.Client {
-  public async forciblyGenerateFromText(
-    givenRequest: GenerateFromTextRequest,
-  ): Promise<GenerateFromTextResponse> {
-    const derivedBatchedRequestBody = toShortfin.BatchedRequestBody([
-      givenRequest.textToImageRequestBody,
-    ]);
-
-    const textToImageSDXLShortfinClient = new Shortfin.TextToImage.SDXL.Client(this.origin, this.headers);
-
-    const outcomeOfGeneratingImage = await textToImageSDXLShortfinClient.generateImageFrom(derivedBatchedRequestBody);
-    const generatedImage = outcomeOfGeneratingImage.forciblyUnwrap(/* matches error propagation of actual StabilityAI Client */);
-
-    const soleGeneratedArtifact: StabilityAI_TextToImage_Pipeline_Output = {
-      base64      : generatedImage.toString(),
-      finishReason: 'SUCCESS',
-      seed        : givenRequest.textToImageRequestBody.seed,
-    };
-
-    return {
-      headers: {},
-      result : {
-        artifacts: [
-          soleGeneratedArtifact,
-        ],
-      },
-    };
-  }
-}
+import {
+  ShimmedStabilityAIClient_Version1_Image,
+} from './Version1/Image';
 
 class ShimmedStabilityAIClient_Version1
   extends HTTP.Client {
